requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
imagehash>=4.3.0
scikit-image>=0.21.0
Pillow>=10.0.0
tqdm>=4.65.0
